pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dnd-combat-sim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "d20>=1.1.2",
  "pydantic>=2.12",
  "python-frontmatter>=1.1.0",
  "requests>=2.31.0",
  "scipy>=1.11.0",
  "ollama>=0.4.7",
  "openai>=1.0",
  "tenacity>=8.3",
  "textual>=0.40",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
